+++ src/main/java/hu/tamas/university/repository/AnswersToPollRepository.java
@@ -3,7 +3,7 @@ package hu.tamas.university.repository;
 import hu.tamas.university.entity.AnswersToPoll;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AnswersToPollRepository extends JpaRepository<AnswersToPoll, Long> {
+public interface AnswersToPollRepository extends JpaRepository<AnswersToPoll, Integer> {
 
 	AnswersToPoll findAnswersToPollById(int id);
 }
